Refactored HealthNews to remote package
