+++ app/src/main/java/com/app/fitspace/data/remote/HealthNewsApiClient.kt
@@ -1,6 +1,5 @@
 package com.app.fitspace.data.remote
 
-import com.app.fitspace.data.local.HealthNews
 import com.google.gson.Gson
 import okhttp3.Callback
 import okhttp3.OkHttpClient
