+++ app/src/main/java/com/app/fitspace/data/remote/HealthNewsResponse.kt
@@ -1,7 +1,5 @@
 package com.app.fitspace.data.remote
 
-import com.app.fitspace.data.local.HealthNews
-
 data class HealthNewsResponse(
     val articles: List<HealthNews>
 )
