+++ app/src/main/java/com/app/fitspace/presentation/fragment/NewsFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.app.fitspace.data.local.HealthNews
+import com.app.fitspace.data.remote.HealthNews
 import com.app.fitspace.data.remote.HealthNewsApiClient
 import com.app.fitspace.databinding.FragmentNewsBinding
 import com.app.fitspace.presentation.adapter.NewsAdapter
